apps/ari-controller/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
requests==2.32.3
websocket-client==1.8.0
